gralloc: Replace USE_GRALLOC1 macro with export_dir switching.
Instead of ensuring all modules using these headers compile with or
without -DUSE_GRALLOC1, change the exported dirs to include the right
version directly. With this change any module can safely use
display_intf_headers without including gralloc_defaults in their
defaults list, or defining the preprocessor macro themselves in an
Android.mk.
